Update Week #02 - Sockets Connection, HTTPS handling.c
Dont make the changes just yet, lets get on call eventually and make the changes then. I have not changed the functions itself, but i have made changes to the core layout of the program. So, the functions u made still do what u had made them to do, but there are other things in those same functions. This version of the program so far can do the following.... I dont wanna explain it.... 2 quiz to prepare for.